+++ packages/xdl/src/UrlUtils.ts
@@ -422,37 +422,6 @@ export function stripJSExtension(entryPoint: string): string {
   return entryPoint.replace(/\.js$/, '');
 }
 
-export function randomIdentifier(length: number = 6): string {
-  const alphabet = '23456789qwertyuipasdfghjkzxcvbnm';
-  let result = '';
-  for (let i = 0; i < length; i++) {
-    const j = Math.floor(Math.random() * alphabet.length);
-    const c = alphabet.substr(j, 1);
-    result += c;
-  }
-  return result;
-}
-
-export function sevenDigitIdentifier(): string {
-  return `${randomIdentifier(3)}-${randomIdentifier(4)}`;
-}
-
-export function randomIdentifierForUser(username: string): string {
-  return `${username}-${randomIdentifier(3)}-${randomIdentifier(2)}`;
-}
-
-export function someRandomness(): string {
-  return [randomIdentifier(2), randomIdentifier(3)].join('-');
-}
-
-export function domainify(s: string): string {
-  return s
-    .toLowerCase()
-    .replace(/[^a-z0-9-]/g, '-')
-    .replace(/^-+/, '')
-    .replace(/-+$/, '');
-}
-
 export function isHttps(urlString: string): boolean {
   return isURL(urlString, { protocols: ['https'] });
 }