+++ packages/xdl/src/__tests__/Project-publishAsync-test.ts
@@ -78,7 +78,7 @@ jest.mock('axios', () => ({
         return mockApiV2Response(
           {
             url: 'https://test.exp.host/@testing/publish-test-app',
-            projectPageUrl: 'https://test.expo.io/@testing/projects/publish-test-app',
+            projectPageUrl: 'https://test.expo.dev/@testing/projects/publish-test-app',
             ids: ['1', '2'],
           },
           options
@@ -124,7 +124,7 @@ describe('publishAsync', () => {
     process.env.EXPO_USE_DEV_SERVER = 'true';
     const result = await publishAsync(projectRoot, { resetCache: true });
     expect(result.url).toBe('https://test.exp.host/@testing/publish-test-app');
-    expect(result.projectPageUrl).toBe('https://test.expo.io/@testing/projects/publish-test-app');
+    expect(result.projectPageUrl).toBe('https://test.expo.dev/@testing/projects/publish-test-app');
 
     process.env.EXPO_USE_DEV_SERVER = 'false';
     const resultOld = await publishAsync(projectRoot, { resetCache: true });