+++ packages/xdl/src/start/startLegacyReactNativeServerAsync.ts
@@ -154,10 +154,13 @@ export async function startReactNativeServerAsync({
   };
 
   if (options.nonPersistent && !Versions.gteSdkVersion(exp, '33.0.0')) {
+    // Expo SDK -32 | React Native -57
     packagerOpts.nonPersistent = true;
   }
 
   if (!Versions.lteSdkVersion(exp, '32.0.0')) {
+    // Expo SDK +33 | React Native +59.8 (hooks): Add asset plugins
+
     // starting with SDK 37, we bundle this plugin with the expo-asset package instead of expo,
     // so check there first and fall back to expo if we can't find it in expo-asset
     packagerOpts.assetPlugins = resolveFrom.silent(projectRoot, 'expo-asset/tools/hashAssetFiles');
@@ -176,6 +179,7 @@ export async function startReactNativeServerAsync({
   }
 
   if (Versions.lteSdkVersion(exp, '15.0.0')) {
+    // Expo SDK -15 | React Native -42: customLogReporterPath is not supported
     delete packagerOpts.customLogReporterPath;
   }
   const userPackagerOpts = exp.packagerOpts;